% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_csv.R
\name{dl_csv}
\alias{dl_csv}
\title{Downloading the valid grids}
\usage{
dl_csv()
}
\description{
This function is required to download the valid grids from the distant folder on Zenodo. 
This way, there is no need to run the Rcpp files "row_generator.cpp", "grid_generator.cpp",
"bigger_grid_generator.cpp".
}
