% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_by_difficulty.R
\name{hide_by_difficulty}
\alias{hide_by_difficulty}
\title{Difficulty Level setup}
\usage{
hide_by_difficulty(lvl, grid)
}
\description{
Hides cells in the Takuzu grid based on the specified difficulty level.
}
